# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rust API",
    "description": "This is the template for the Rust API",
    "contact": {
      "name": "Batleforc",
      "url": "https://weebo.fr",
      "email": "maxleriche.60@gmail.com"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/otp/activate": {
      "get": {
        "tags": [
          "Auth>Otp"
        ],
        "summary": "Start the totp activate process",
        "description": "Start the totp activate process",
        "operationId": "generate",
        "parameters": [
          {
            "name": "Authorization-type",
            "in": "header",
            "description": "Type de token (oidc ou buildin)",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AuthType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "QrCode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenOtp"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "deprecated": false,
        "security": [
          {
            "access_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "Auth>Otp"
        ],
        "summary": "End the totp activate process",
        "description": "End the totp activate process",
        "operationId": "activate",
        "parameters": [
          {
            "name": "Authorization-type",
            "in": "header",
            "description": "Type de token (oidc ou buildin)",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AuthType"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateOtp"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "deprecated": false,
        "security": [
          {
            "access_token": []
          }
        ]
      }
    },
...
    "/api/user/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get one by uid user",
        "description": "Get one by uid user\n\nGet one user by id",
        "operationId": "getoneuser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id de l'utilisateur",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Authorization-type",
            "in": "header",
            "description": "Type de token (oidc ou buildin)",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AuthType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUser"
                }
              }
            }
          },
          "400": {
            "description": "Error message"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "deprecated": false,
        "security": [
          {
            "access_token": []
          },
          {
            "oidc": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivateOtp": {
        "type": "object",
        "required": [
          "otp_code"
        ],
        "properties": {
          "otp_code": {
            "type": "string"
          }
        }
      },
...
      "AuthType": {
        "type": "string",
        "enum": [
          "Oidc",
          "BuildIn"
        ]
      },
...
      "GenOtp": {
        "type": "object",
        "required": [
          "url",
          "qr_code"
        ],
        "properties": {
          "qr_code": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
...
      "PublicUser": {
        "type": "object",
        "required": [
          "id",
          "nom",
          "prenom"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "access_token": {
        "type": "http",
        "scheme": "bearer"
      },
      "oidc": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://localhost:8080/oauth/v2/authorize",
            "tokenUrl": "http://localhost:8080/oauth/v2/token",
            "scopes": {
              "email": "Access email (needed)",
              "offline_access": "Access offline (needed)",
              "openid": "Access auth (needed)",
              "profile": "Access profile (needed)"
            }
          }
        },
        "description": "Zitadel dev"
      },
      "refresh_token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}